backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# ML Libraries (optional for now, needed later)
# torch==2.1.2
# transformers==4.37.0
# accelerate==0.26.1

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
